[Builder] Move shared build-pod spec-attribute resolver to base
buildah imported from kaniko only to reuse the runtime->pod-spec-attribute
resolver, which is a layering smell (both backends should depend on base, not
each other). Fold the resolver and its apply-loop into a single engine-neutral
base.resolve_build_pod_spec_attributes(), and have both Kaniko and Buildah call
it. Removes the buildah->kaniko import and the duplicated loop; behaviour is
unchanged (191 builder unit tests pass).
